services/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sentence-transformers==2.2.2
pandas==2.2.2
pyarrow==16.1.0
sqlalchemy==2.0.23
alembic==1.13.0
plaid-python==21.0.0
python-dotenv==1.0.0
schedule==1.2.0
yfinance==0.2.28
numpy==1.26.4
alpaca-trade-api==3.2.0
websockets>=10.0
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0
reportlab==4.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.31.0
celery==5.3.6
redis==5.0.1
flower==2.0.1
python-dateutil==2.8.2
scikit-learn==1.3.2
joblib==1.3.2
psutil==7.0.0
python-json-logger==3.3.0